Waste-Managment-main/Waste-Managment-main/smart-waste-sorter/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
pymongo==4.6.1
motor==3.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
fpdf2==2.7.6
pandas==2.1.3
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0
